fix : new mission page
