+++ client/src/App.js
@@ -6,6 +6,7 @@ import Header from "./components/Header";
 import SignUp from "./pages/SignUp";
 import "bootstrap/dist/css/bootstrap.css";
 import MissionDetail from "./pages/MissionDetail";
+import NewMission from "./pages/NewMission";
 
 function App() {
   return (
@@ -14,14 +15,8 @@ function App() {
       <Routes>
         <Route path="/" element={<Landing />}></Route>
         <Route path="/mypage" element={<Mypage />}></Route>
-        <Route
-          path={"/missiondetail"}
-          element={<MissionDetail isWriting={false} />}
-        />
-        <Route
-          path="/newmission"
-          element={<MissionDetail isWriting={true} />}
-        />
+        <Route path={"/missiondetail"} element={<MissionDetail />} />
+        <Route path="/newmission" element={<NewMission />} />
         <Route path="signup" element={<SignUp />}></Route>
       </Routes>
     </div>
